Implemented Firebase for Login/Logout

+++ src/hooks/useDeleteRobot.js
@@ -1,15 +1,20 @@
 import { useCallback } from "react";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { deleteRobot } from "../store/robotSlice";
-import { BASE_URL } from "../utils/api";
+import { BASE_URL, getAuthHeaders } from "../utils/api";
 
 const useDeleteRobot = () => {
   const dispatch = useDispatch();
+  const user = useSelector((store) => store.user.currentUser);
 
   const removeRobot = useCallback(async (robotId) => {
+    if (!user?.uid) return;
+
     try {
+      const authHeaders = await getAuthHeaders();
       const response = await fetch(`${BASE_URL}/api/robots/${robotId}`, {
         method: "DELETE",
+        headers: authHeaders,
       });
 
       if (!response.ok) {
@@ -20,7 +25,7 @@ const useDeleteRobot = () => {
     } catch (error) {
       console.error("Failed to delete robot:", error);
     }
-  }, [dispatch]);
+  }, [dispatch, user?.uid]);
 
   return removeRobot;
 };
